# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NPI Website API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme."
      }
    },
...
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate user and get access token",
        "description": "Sends credentials to Tesla Factory API to validate user and return a JWT token. The token is required for subsequent API calls. Rate limited to 5 attempts per minute.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@tesla.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Successfully authenticated and redirected to home page",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "session=encrypted_session_data; HttpOnly; Secure; SameSite=Lax; Path=/"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid username or password"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/factory/thing/move-to-process": {
      "post": {
        "summary": "Move a thing to a specific process step",
        "description": "Moves a thing to a specified process, flow, and flow step",
        "tags": ["Factory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "thingName",
                  "processid",
                  "flowid",
                  "flowstepid",
                  "token",
                  "username"
                ],
                "properties": {
                  "thingName": {
                    "type": "string",
                    "description": "Name of the thing to move",
                    "example": "THING001"
                  },
                  "processid": {
                    "type": "string",
                    "description": "Process ID to move the thing to",
                    "example": "PROCESS_001"
                  },
                  "flowid": {
                    "type": "string",
                    "description": "Flow ID within the process",
                    "example": "FLOW_001"
                  },
                  "flowstepid": {
                    "type": "string",
                    "description": "Flow step ID within the flow",
                    "example": "STEP_001"
                  },
                  "token": {
                    "type": "string",
                    "description": "JWT authentication token"
                  },
                  "username": {
                    "type": "string",
                    "description": "Username of the actor performing the action",
                    "example": "user@tesla.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thing moved to process successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Thing entered process successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}